Code review changes: - Move FileAttachmentForm.spec.tsx.snap to public - Split InferDomainResponse.ts - Move inferDomainFromFile() & getServerFilePreview() to a new file utils.js in internal/componentes/assays. Keep InferDomainResponse class and move this file to public. - Update index.ts. - Update public docs.
