+++ packages/components/src/internal/components/assay/utils.ts
@@ -1,34 +1,6 @@
-import { fromJS, List, Record } from 'immutable';
-import { buildURL, QueryColumn } from '..';
-import { Ajax, Utils } from '@labkey/api';
-
-export class InferDomainResponse extends Record({
-    data: List<any>(),
-    fields: List<QueryColumn>(),
-}) {
-    data: List<any>;
-    fields: List<QueryColumn>;
-
-    static create(rawModel): InferDomainResponse {
-        let data = List<any>();
-        let fields = List<QueryColumn>();
-
-        if (rawModel) {
-            if (rawModel.data) {
-                data = fromJS(rawModel.data);
-            }
-
-            if (rawModel.fields) {
-                fields = List(rawModel.fields.map(field => QueryColumn.create(field)));
-            }
-        }
-
-        return new InferDomainResponse({
-            data,
-            fields,
-        });
-    }
-}
+import {Ajax, Utils} from "@labkey/api";
+import {buildURL} from "../../url/AppURL";
+import {InferDomainResponse} from "../../../public/InferDomainResponse";
 
 export function inferDomainFromFile(file: File, numLinesToInclude: number): Promise<InferDomainResponse> {
     return new Promise((resolve, reject) => {
